requirements.txt
streamlit
fastapi
uvicorn
numpy
joblib
scikit-learn
tensorflow-cpu
pydantic
pandas
matplotlib
seaborn
